Move NUSMods API into global singleton
So that we only run setup code once, even if we have multiple AYs.

This prevents the scraper from trying to create multiple
schools/AYs/semesters.  Okay, this is a bit of a lie, because if we have
multiple AYs the school setup code will run multiple times. This should
be fine though, since we should only be setting up schools extremely
rarely.

This also allows all GraphQLPersist instances to share the same queue,
which is nice as every task has its own persist instance.